fix: complete status omit start time
