+++ src/utils/config/index.ts
@@ -33,9 +33,17 @@ export const DEFAULT_APP_ACTOR_NAME = "approveman[bot]";
  */
 export const getAppActorName = (): string => {
   if (process.env.GHE_HOST) {
+    /* istanbul ignore else */
     if (process.env.APP_ACTOR_NAME_OVERRIDE) {
       return process.env.APP_ACTOR_NAME_OVERRIDE;
     } else {
+      /**
+       * Since there is already a static health check at
+       * the beginning of the server boot up. This branch
+       * will never run in practice. However, it's left
+       * here for some unknown situations since the host
+       * environment can get weird sometimes.
+       */
       throw Error(GITHUB_ENTERPRISE_APP_ACTOR_NOT_FOUND);
     }
   }
